Update delete dialog confirm fragment
